Ensure original_date cannot be null
This commit adds a database validation to ensure that an original_date
must always be filled. This is important because we'll use this date for
ordering moments when we're displaying them.

I've also added a model validation and a before hook to make sure we
default the date to the current time if it isn't explicitly provided.