Allow setting original date for moments in the UI
This commit updates the MomentForm component to include a new Datepicker
for setting `original_date` on moments. The Datepicker component makes
use of Airbnb's `react-dates` component to render a calendar.

The `original_date` for a moment must be either the present date or a
date in the past - you cannot log a moment for the future.