Remove webpack and add js / css bundling with tailwind (#55)
## What

This PR removes webpack and replaces it with js and css bundling for Rails and with Tailwind. It breaks all of the styling but that will get updated in a followup PR.

I've also added action text for moments which will now use a `content` column. Existing moments have content stored on the `body` field, but that will get moved to content in a followup PR.